Some more updates
Hey guys,

Did some more updates-- mostly tinkering with cosmetics. Structure and
cleanliness to code is essential if you don't want to get lost in all
your stuff and others will need to work on it. I briefly looked at the
objectdatabase; could we not simply stick the cells with provenance data
as they are read from the Excel sheets in a single text-field per JSON
item. and then index on that text? That way we would not necessarily
need to do any mapping of people, places etc. Just a thought.