CSS changes-- everything still works! :)
Done some CSS styling; everything still works :) Some known bugs:
- resizing screen will make the navigation menu difficult to use;
- each time a button is pressed in the navigation menu the whole page
reloads-- I now have it load without posting and it still has this
behavior. Therefore, this is not because of the $.post requests to the
server from the actions, but possibly because every click follows the
page to "/", which then reloads the whole page. What we want is a static
'index' page with a static menu bar, of which only the rest of the page
is dynamic. Any ideas?
- uploading results will need to be rendered in the modal box once a
file has been uploaded;
- do we still need to indicate which column contains what? If so,
perhaps restyle that?
- maybe we can have the server return the results of reading the XLSX
sheet, indicate the different columns, so that the user can indicate
which column is what;
- userprofile: users will need to be able to see what they have uploaded
and manage that in some sense (edit, delete, upload a small thumbnail
image etc.)
- searching: pressing enter too quickly following another press will
show up different results;
- toying with animations for search results-- we'll need a hoverover so
that info will show up of these items...