Updates on profile
Despite Abe's suggestion, I think it makes sense to have the profile as
well in a modal box-- uniformity is key to make sure the wrinkleys
aren't getting confused.  I've stuck the Profile bit in a modal box and
put a tab system in place which I think just requires a bit of styling
and additional functionality, but then most of that should fine-- we'll
need to add a little functionality so that uploaded data can be viewed
in a list. We don't necessarily need to keep a copy of the XLSX file on
the server, but should at least keep a record of it and all the objects
that came from it. For Abe's search within uploaded data, the user will
search for one object to manage it's data so we can integrate that into
this box as not many results will be returned. The only bit of extra
functionality we'll need to add on that end is that the object can be
edited once it pops up in the search results