Tweaks to markup tool
I know this is optional, but I'm just curious to see if we can get it to
work. I think we just need to change a few things on routes.js and
objectdatabase.js to get it to work-- anyone want to have a quick look?

Also, ignore the test.ejs and test.js for now-- feel free to // comment
them out whereever necessary; I want to experiment a little with them.