Question for the MongoDB wizkids
Ok, I think it's almost working-- if this works we can simply copy/paste
it for locations and times. Could someone with better knowledge on
MongoDB and its searching functions have a look at how to query all
documents in the persons, locations, and times collections for the given
term? Would we need separate functions?